- DEF-803: Escape HTTP URLs
* http_client now expects a properly formatted URI

* Remove escape/unescape from http_client, merged into
  uri/Encode and Decode functions

* Remove URI decoding in FileHandler since URI() needs a properly
  formatted URI, which it should already be in the requst line
