+++ engine/dlib/src/dlib/http_client.cpp
@@ -818,9 +818,7 @@ bail:
     {
         dmSocket::Result sock_res;
 
-        char encoded_path[DMPATH_MAX_PATH];
-        dmURI::Encode(path, encoded_path, DMPATH_MAX_PATH);
-        sock_res = SendRequest(client, &response, encoded_path, method);
+        sock_res = SendRequest(client, &response, path, method);
 
         if (sock_res != dmSocket::RESULT_OK)
         {
@@ -1039,93 +1037,6 @@ bail:
     dmHttpCache::HCache GetHttpCache(HClient client)
     {
         return client->m_HttpCache;
-    }
-
-    // http://en.wikipedia.org/wiki/Percent-encoding
-    static bool IsUnreserved(char c)
-    {
-        if (c >= 'a' && c <= 'z') {
-            return true;
-        } else if (c >= 'A' && c <= 'Z') {
-            return true;
-        } else if (c >= '0' && c <= '9') {
-            return true;
-        } else {
-            switch (c) {
-            case '-':
-            case '_':
-            case '.':
-            case '~':
-                return true;
-            }
-        }
-        return false;
-    }
-
-    Result Escape(const char* src, char* dst, uint32_t dst_len)
-    {
-        assert(src != (const char*) dst);
-        assert(dst_len > 0);
-
-        Result r = RESULT_OK;
-        uint32_t dst_left = dst_len - 1; // Make room for null termination
-        const char* s = src;
-        char* d = dst;
-        while (*s) {
-            char c = *s;
-            if (IsUnreserved(c)) {
-                if (dst_left >= 1) {
-                    *d = c;
-                    s++;
-                    d++;
-                    dst_left--;
-                } else {
-                    r = RESULT_INVAL;
-                    break;
-                }
-            } else {
-                if (dst_left >= 3) {
-                    DM_SNPRINTF(d, 4, "%%%02X", c);
-                    s++;
-                    d += 3;
-                    dst_left -= 3;
-                } else {
-                    r = RESULT_INVAL;
-                    break;
-                }
-            }
-        }
-
-        *d = '\0';
-
-        return r;
-    }
-
-    void Unescape(const char* src, char* dst)
-    {
-        size_t left = strlen(src);
-        while (left > 0) {
-            if (src[0] == '+') {
-                *dst = ' ';
-                src++;
-                left--;
-            }
-            else if (left > 2 && src[0] == '%' && isxdigit(src[1]) && isxdigit(src[2])) {
-                char tmp[3];
-                tmp[0] = src[1];
-                tmp[1] = src[2];
-                tmp[2] = '\0';
-                *dst = (char) strtoul(tmp, 0, 16);
-                src += 3;
-                left -= 3;
-            } else {
-                *dst = *src;
-                src++;
-                left--;
-            }
-            dst++;
-        }
-        *dst = '\0';
     }
 
 #undef HTTP_CLIENT_SENDALL_AND_BAIL
